\name{fixCountryCol}
\alias{fixCountryCol}
\title{Standardize a vector of countries}
\usage{
fixCountryCol(inVec)
}
\arguments{
\item{inVec}{A vector of countries that are unstandardized}
}
\description{
This function makes the following corrections:
 \cr
  'United Kingdom' <- 'UK' \cr
  "United States" <- "USA" \cr
  "The United States" <- "USA" \cr
  "Untied States" <- "USA" \cr
  "United Staets" <- "USA" \cr
  "The United States of America" <- "USA" \cr
  "United Arab Emirates" <- "UAE" \cr
  "Iran, Islamic Republic of" <- "Iran" \cr
  "Moldova, Republic of" <- "Moldova" \cr
  "The Former Yugoslav Republic of Macedonia" <- "FYR Macedonia" \cr
  "Union of Myanmar (republic of the)" <- "Myanmar" \cr
  "Afghanistan, Islamic Republic of" <- "Afghanistan" \cr
  "Syrian Arab Republic" <- "Syria" \cr
  "Korea, Republic of" <- "Korea (Rep. of)" \cr
  "Cote D'ivoire" <- "Cote d'Ivoire" \cr
  "Macao, China" <- "Macau, China" \cr
  "The Former Yugoslav Republic of Macedonia" <- "FYR Macedonia" \cr
  "Afghanistan, Islamic Republic of" <- "Afghanistan" \cr
  "United Arab Emirates" <- "UAE" \cr
  "Syrian Arab Republic" <- "Syria" \cr
  "Korea, Republic of" <- "Korea (Rep. of)" \cr
  "Democratic Republic of the Congo" <- "Congo (Dem. Rep. of)" \cr
  "Congo, Republic of the" <- "Congo (Rep. of)" \cr
  "Ivory Coast" <- "Cote d'Ivoire" \cr
  "Cote D'ivoire" <- "Cote d'Ivoire" \cr
  "Macao, China" <- "Macau, China" \cr
  "Iran, Islamic Republic of" <- "Iran" \cr
  "Us Virgin Islands" <- "US Virgin Islands" \cr
  "Moldova, Republic of" <- "Moldova" \cr
  "The Former Yugoslav Republic of Macedonia" <- "FYR Macedonia" \cr
  "Afghanistan, Islamic Republic of" <- "Afghanistan" \cr
  "United Arab Emirates" <- "UAE" \cr
  "Syrian Arab Republic" <- "Syria" \cr
  "Korea, Republic of" <- "Korea (Rep. of)" \cr
  "Democratic Republic of the Congo" <- "Congo (Dem. Rep. of)" \cr
  "Congo, Republic of the" <- "Congo (Rep. of)" \cr
  "Macao, China" <- "Macau, China" \cr
}




